Added full support for Events!
Added all 13 events tracks. Support for the 4 states (off, on, flash, fade), and both colors.

Left Click to place RED blocks, right click to place BLUE blocks.

Press 'A' + the color to place an FLASH block for that color.

Press 'D' + the color to place a FADE block for that color.

Press 'W' or middle click to place an OFF block for either color.

I don't know if these are all going to work, but we will see!